New WIP apriltag driving function, need to tune the new PID controllers
Two useful functions exist in apriltags now, both of which output a ChassisSpeeds object which will be fed to the SwerveDrive. One function allows the robot to follow the nearest apriltag at a desired distance away while the other allows the robot to move to a desired position. The main thing left to do is to tune the ProfiledPIDControllers which output the ChassisSpeeds components. Both trapezoid constraints and PID constants are necessary.